refactor: extract HeroPacingGraph into its own file
